one step to fin

+++ resources/item.py
@@ -1,6 +1,7 @@
 from flask import request
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
+from flask_jwt_extended import jwt_required, get_jwt # resource protection
 
 from schemas import ItemSchema, ItemUpdateSchema
 from db import db
@@ -9,9 +10,12 @@ from sqlalchemy.exc import SQLAlchemyError
 
 blp = Blueprint("Items", "items", description="Operations on items")
 
+# for demonstration purpose, only protecting Item resource to login users
 @blp.route("/item/<string:item_id>")
 class Item(MethodView):
-    
+    # route marked with jwt_required would have request header Content-Type: Authorization and application/json: Bearer <jwt>
+    # the jwt is returned from user /login route
+    @jwt_required() 
     @blp.response(200, ItemSchema) # with this decorator, validate as ItemSchema, documentation updated and wrap response 200
     def get(self, item_id):
         # try:
@@ -22,25 +26,32 @@ class Item(MethodView):
         item = ItemModel.query.get_or_404(item_id) 
         return item 
         
-
+    @jwt_required()
     @blp.response(204)
     def delete(self, item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"): # field stored in the app.py:add_claims_to_jwt()
+            abort(401, message="Admin privilege required to delete items.")
+        
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item) # DELETE FROM Items WHERE id = item_id
         db.session.commit()
         return {"message" : "Item deleted."}
 
+    @jwt_required(fresh=True)
     @blp.arguments(ItemUpdateSchema) # register to the update schema 
     @blp.response(200, ItemSchema)
     def put(self, item_data, item_id): # item_data the injected data always come before url argument <item_id>
         # uses update-or-insert as if in the raw dictionary
         item = ItemModel.query.get_or_404(item_id)
         if item:
-            item.price = item_data['price']
-            item.name = item_data['name']
+            if 'price' in item_data: # field-wise validation, replaced by more concise code if item_data has k fields
+                item.price = item_data['price']
+            if 'name' in item_data:
+                item.name = item_data['name']
         else:
             item = ItemModel(id=item_id, **item_data)
-        
+            
         db.session.add(item)
         db.session.commit()
         
@@ -48,10 +59,12 @@ class Item(MethodView):
 
 @blp.route("/item")
 class ItemList(MethodView):
+    @jwt_required() 
     @blp.response(200, ItemSchema(many=True)) # return multiple instances to be validated
     def get(self):
         return ItemModel.query.all() # this queries all items in the Items table, equiv to SELECT * FROM Items
 
+    @jwt_required(fresh=True)
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema) # 201 signifies POST request success
     def post(self, item_data):
